requirements_backup.txt
streamlit
anthropic
python-dotenv
PyPDF2
python-docx
pandas
plotly
reportlab
Pillow